chore(engine-v2): Clean-up solve module
