+++ engine/crates/engine-v2/src/execution/state.rs
@@ -47,7 +47,7 @@ impl<'ctx, R: Runtime> OperationExecutionState<'ctx, R> {
     pub(super) fn new(ctx: ExecutionContext<'ctx, R>) -> Self {
         Self {
             ctx,
-            response_object_sets: vec![None; ctx.operation.solution.response_object_set_definitions.len()],
+            response_object_sets: vec![None; ctx.operation.cached.solved.response_object_set_definitions.len()],
             plan_to_parent_count: ctx.operation.plan.plans.iter().map(|plan| plan.parent_count).collect(),
             response_modifier_to_parent_count: ctx
                 .operation
