+++ engine/crates/engine-v2/src/operation/solve/model/argument.rs
@@ -8,7 +8,7 @@ impl<'a> FieldArgument<'a> {
     /// Used for GraphQL query generation to only include values in the query string that would be
     /// present after query sanitization.
     pub(crate) fn value_as_sanitized_query_const_value_str(&self) -> Option<&'a str> {
-        Some(match &self.ctx.operation_solution.query_input_values[self.value_id] {
+        Some(match &self.ctx.operation.query_input_values[self.value_id] {
             QueryInputValueRecord::EnumValue(id) => self.ctx.schema.walk(*id).name(),
             QueryInputValueRecord::Boolean(b) => {
                 if *b {
