+++ engine/crates/engine-v2/src/operation/solve/solver/shapes/partition.rs
