+++ engine/crates/engine-v2/src/resolver/graphql/deserialize/errors.rs
@@ -20,7 +20,7 @@ impl<'resp> RootGraphqlErrors<'resp> {
     pub fn new<R: Runtime>(ctx: &ExecutionContext<'resp, R>, response: SubgraphResponseRefMut<'resp>) -> Self {
         Self {
             response,
-            response_keys: &ctx.operation.solution.response_keys,
+            response_keys: &ctx.operation.cached.solved.response_keys,
         }
     }
 }
